# Faster Deflate