--- a/src/ImageSharp/Formats/Png/Zlib/DeflaterEngine.cs
+++ b/src/ImageSharp/Formats/Png/Zlib/DeflaterEngine.cs
@@ -525,12 +525,15 @@ namespace SixLabors.ImageSharp.Formats.Png.Zlib
                         break;
 
                     case 2:
-                        if (pinnedWindow[++scan] == pinnedWindow[++match]
-                            && pinnedWindow[++scan] == pinnedWindow[++match])
+                        if ((short*)pinnedWindow[++scan] == (short*)pinnedWindow[++match])
                         {
+                            ++scan;
+                            ++match;
                             break;
                         }
 
+                        ++scan;
+                        ++match;
                         break;
 
                     case 3:
@@ -544,14 +547,15 @@ namespace SixLabors.ImageSharp.Formats.Png.Zlib
                         break;
 
                     case 4:
-                        if (pinnedWindow[++scan] == pinnedWindow[++match]
-                            && pinnedWindow[++scan] == pinnedWindow[++match]
-                            && pinnedWindow[++scan] == pinnedWindow[++match]
-                            && pinnedWindow[++scan] == pinnedWindow[++match])
+                        if ((int*)pinnedWindow[++scan] == (int*)pinnedWindow[++match])
                         {
+                            scan += 3;
+                            match += 3;
                             break;
                         }
 
+                        scan += 3;
+                        match += 3;
                         break;
 
                     case 5: