--- a/src/ImageSharp/Formats/Png/Zlib/DeflaterHuffman.cs
+++ b/src/ImageSharp/Formats/Png/Zlib/DeflaterHuffman.cs
@@ -36,11 +36,6 @@ namespace SixLabors.ImageSharp.Formats.Png.Zlib
 
         private const int EofSymbol = 256;
 
-        private static readonly short[] StaticLCodes;
-        private static readonly byte[] StaticLLength;
-        private static readonly short[] StaticDCodes;
-        private static readonly byte[] StaticDLength;
-
         private Tree literalTree;
         private Tree distTree;
         private Tree blTree;
@@ -57,49 +52,6 @@ namespace SixLabors.ImageSharp.Formats.Png.Zlib
         private int lastLiteral;
         private int extraBits;
         private bool isDisposed;
-
-        // TODO: These should be pre-generated array/readonlyspans.
-        static DeflaterHuffman()
-        {
-            // See RFC 1951 3.2.6
-            // Literal codes
-            StaticLCodes = new short[LiteralNumber];
-            StaticLLength = new byte[LiteralNumber];
-
-            int i = 0;
-            while (i < 144)
-            {
-                StaticLCodes[i] = BitReverse((0x030 + i) << 8);
-                StaticLLength[i++] = 8;
-            }
-
-            while (i < 256)
-            {
-                StaticLCodes[i] = BitReverse((0x190 - 144 + i) << 7);
-                StaticLLength[i++] = 9;
-            }
-
-            while (i < 280)
-            {
-                StaticLCodes[i] = BitReverse((0x000 - 256 + i) << 9);
-                StaticLLength[i++] = 7;
-            }
-
-            while (i < LiteralNumber)
-            {
-                StaticLCodes[i] = BitReverse((0x0c0 - 280 + i) << 8);
-                StaticLLength[i++] = 8;
-            }
-
-            // Distance codes
-            StaticDCodes = new short[DistanceNumber];
-            StaticDLength = new byte[DistanceNumber];
-            for (i = 0; i < DistanceNumber; i++)
-            {
-                StaticDCodes[i] = BitReverse(i << 11);
-                StaticDLength[i] = 5;
-            }
-        }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DeflaterHuffman"/> class.
@@ -122,12 +74,77 @@ namespace SixLabors.ImageSharp.Formats.Png.Zlib
             this.pinnedLiteralBuffer = (short*)this.literalBufferHandle.Pointer;
         }
 
+        // See RFC 1951 3.2.6
+        // Literal codes
+        private static short[] StaticLCodes => new short[]
+        {
+            12, 140, 76, 204, 44, 172, 108, 236, 28, 156, 92, 220, 60, 188, 124, 252,
+            2, 130, 66, 194, 34, 162, 98, 226, 18, 146, 82, 210, 50, 178, 114, 242,
+            10, 138, 74, 202, 42, 170, 106, 234, 26, 154, 90, 218, 58, 186, 122, 250,
+            6, 134, 70, 198, 38, 166, 102, 230, 22, 150, 86, 214, 54, 182, 118, 246,
+            14, 142, 78, 206, 46, 174, 110, 238, 30, 158, 94, 222, 62, 190, 126, 254,
+            1, 129, 65, 193, 33, 161, 97, 225, 17, 145, 81, 209, 49, 177, 113, 241, 9,
+            137, 73, 201, 41, 169, 105, 233, 25, 153, 89, 217, 57, 185, 121, 249, 5,
+            133, 69, 197, 37, 165, 101, 229, 21, 149, 85, 213, 53, 181, 117, 245, 13,
+            141, 77, 205, 45, 173, 109, 237, 29, 157, 93, 221, 61, 189, 125, 253, 19,
+            275, 147, 403, 83, 339, 211, 467, 51, 307, 179, 435, 115, 371, 243, 499,
+            11, 267, 139, 395, 75, 331, 203, 459, 43, 299, 171, 427, 107, 363, 235, 491,
+            27, 283, 155, 411, 91, 347, 219, 475, 59, 315, 187, 443, 123, 379, 251, 507,
+            7, 263, 135, 391, 71, 327, 199, 455, 39, 295, 167, 423, 103, 359, 231, 487,
+            23, 279, 151, 407, 87, 343, 215, 471, 55, 311, 183, 439, 119, 375, 247, 503,
+            15, 271, 143, 399, 79, 335, 207, 463, 47, 303, 175, 431, 111, 367, 239, 495,
+            31, 287, 159, 415, 95, 351, 223, 479, 63, 319, 191, 447, 127, 383, 255, 511,
+            0, 64, 32, 96, 16, 80, 48, 112, 8, 72, 40, 104, 24, 88, 56, 120, 4, 68, 36,
+            100, 20, 84, 52, 116, 3, 131, 67, 195, 35, 163
+        };
+
+        private static ReadOnlySpan<byte> StaticLLength => new byte[]
+        {
+            8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
+            8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
+            8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
+            8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
+            8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
+            8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
+            8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
+            8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
+            8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
+            9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9,
+            9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9,
+            9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9,
+            9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9,
+            9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9,
+            9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9,
+            9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9,
+            7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7,
+            7, 7, 7, 7, 7, 7, 7, 7, 8, 8, 8, 8, 8, 8
+        };
+
+        // Distance codes and lengths.
+        private static short[] StaticDCodes => new short[]
+        {
+            0, 16, 8, 24, 4, 20, 12, 28, 2, 18, 10, 26, 6, 22, 14,
+            30, 1, 17, 9, 25, 5, 21, 13, 29, 3, 19, 11, 27, 7, 23
+        };
+
+        private static ReadOnlySpan<byte> StaticDLength => new byte[]
+        {
+            5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5,
+            5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5
+        };
+
         /// <summary>
         /// Gets the lengths of the bit length codes are sent in order of decreasing probability, to avoid transmitting the lengths for unused bit length codes.
         /// </summary>
-        private static ReadOnlySpan<byte> BitLengthOrder => new byte[] { 16, 17, 18, 0, 8, 7, 9, 6, 10, 5, 11, 4, 12, 3, 13, 2, 14, 1, 15 };
+        private static ReadOnlySpan<byte> BitLengthOrder => new byte[]
+        {
+            16, 17, 18, 0, 8, 7, 9, 6, 10, 5, 11, 4, 12, 3, 13, 2, 14, 1, 15
+        };
 
-        private static ReadOnlySpan<byte> Bit4Reverse => new byte[] { 0, 8, 4, 12, 2, 10, 6, 14, 1, 9, 5, 13, 3, 11, 7, 15 };
+        private static ReadOnlySpan<byte> Bit4Reverse => new byte[]
+        {
+            0, 8, 4, 12, 2, 10, 6, 14, 1, 9, 5, 13, 3, 11, 7, 15
+        };
 
         /// <summary>
         /// Gets the pending buffer to use.