--- a/src/ImageSharp/Formats/Png/Zlib/DeflaterPendingBuffer.cs
+++ b/src/ImageSharp/Formats/Png/Zlib/DeflaterPendingBuffer.cs
@@ -172,8 +172,6 @@ namespace SixLabors.ImageSharp.Formats.Png.Zlib
                 this.bufferMemoryOwner = null;
                 this.isDisposed = true;
             }
-
-            GC.SuppressFinalize(this);
         }
     }
 }