### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Attempting to chip away at the slowness of the managed `ZlibStream`.

I'm in two minds here. Do we continue and run our own version or do we reimplement the wrapper around .NET `DeflateStream` and hope that #1027 is eventually fixed upstream?

We're currently running at 2-2.5x slower than System.Drawing in our benchmarks. I'm managed to save around 10% with these changes.

**I need help optimizing this**. I don't fully understand the deflate process so cannot make any drastic algorithmic improvements. All my changes since copying this from SharpZipLib have been minor and mostly focused around removing bonds checks.

### UPDATE.

Going As-is with this just now. I got the diff down from ~2.5x to the below. Wasn't able to isolate the allocations but can investigate that in the future. 

|               Method |       Runtime | LargeImage |     Mean |    Error |   StdDev | Ratio | RatioSD |    Gen 0 |    Gen 1 |    Gen 2 |  Allocated |
|--------------------- |-------------- |----------- |---------:|---------:|---------:|------:|--------:|---------:|---------:|---------:|-----------:|
| 'System.Drawing Png' |    .NET 4.7.2 |      False | 12.87 ms | 4.162 ms | 0.228 ms |  1.00 |    0.00 | 515.6250 | 234.3750 | 234.3750 | 2079.95 KB |
|     'ImageSharp Png' |    .NET 4.7.2 |      False | 31.00 ms | 4.895 ms | 0.268 ms |  2.41 |    0.06 | 500.0000 | 500.0000 | 500.0000 | 2323.56 KB |
|                      |               |            |          |          |          |       |         |          |          |          |            |
| 'System.Drawing Png' | .NET Core 2.1 |      False | 14.33 ms | 9.995 ms | 0.548 ms |  1.00 |    0.00 | 531.2500 | 250.0000 | 250.0000 | 2076.73 KB |
|     'ImageSharp Png' | .NET Core 2.1 |      False | 30.77 ms | 1.910 ms | 0.105 ms |  2.15 |    0.09 | 562.5000 | 531.2500 | 531.2500 | 2321.25 KB |
|                      |               |            |          |          |          |       |         |          |          |          |            |
| 'System.Drawing Png' | .NET Core 3.1 |      False | 14.08 ms | 7.838 ms | 0.430 ms |  1.00 |    0.00 | 234.3750 | 218.7500 | 218.7500 |  964.29 KB |
|     'ImageSharp Png' | .NET Core 3.1 |      False | 30.79 ms | 0.698 ms | 0.038 ms |  2.19 |    0.07 | 562.5000 | 531.2500 | 531.2500 | 2321.24 KB |

<!-- Thanks for contributing to ImageSharp! -->