@JimBobSquarePants 2020-04-06
>the one in ImageSharp is faster than the BCL one, consistent and not bugged, but just still a bit slower than the one from System.Drawing then

@Sergio0694  No our one is slower than the BCL but not buggy.