@Sergio0694 2020-04-06
Oh, gotcha. Well I have to say I find it really weird to hear about such an important class in the BCL being bugged like that, and left in that state for years. Are there open issues on `dotnet/runtime` about it? Also I wonder then given the situation if it's really worth it to spend all this time trying to optimize this self-contained version, as opposed to trying to bring attention to that bug in the BCL and get someone to fix it, or maybe contribute directly there to get it fixed if possible?

I guess it'd just be bad if they eventually fixed it, having to scrap all the hard work for the implementation in this library that would've just acted as a temporary band-aid then 😄