@JimBobSquarePants 2020-04-08
@Sergio0694 Take your pick.

https://github.com/dotnet/runtime/issues/30539
https://github.com/dotnet/runtime/issues/1549

Someone proposed an API for ZlibStream but again it's a wrapper round the varying implementation and doesn't expose the Compression level.
https://github.com/dotnet/runtime/issues/2236