@Sergio0694 2020-04-08
Wow, I went through the first two issues you linked and man... It might be by design, but it sure looks incredibly inconvenient for devs working on multiple platforms and trying to have fine control 😕

I get that technically the guarantee is that once decompressed you'll always get with the same original data, but I can see how having completely separate implementations under the hood with no real control is probably not what many devs would want to see in their apps.

It just too bad though to have so much time spent into basically rebuilding the same exact API from scratch just to work around this shortcoming in the BCL. If that API proposal for the `DeflateStream` got approved, with also your suggestion for the explicit compression level, it'd be awesome!