@saucecontrol 2020-04-08
I don't think the BCL implementation is buggy so much as it's optimized in ways that may not be ideal for PNG use.  The 'bug' in question is an optimization Intel added to zlib that make it quite a bit faster, with negligible impact to compression for the majority of cases.  There was, however, a regression in compression of sparse data, which is more common in images than in text.  The BCL team has simply been hesitant to give up the speed improvement for the majority to restore the compression for the minority.

That said, even the Intel-modified zlib doesn't represent the best performance available.  Cloudflare and Google have also created zlib variants with various optimizations to performance and compression.  There's an alternate project called [zlib-ng](https://github.com/zlib-ng/zlib-ng) that has gathered all of these improvements together and modernized the zlib codebase, which might serve as a good model for a high-performance managed zlib replacement.  It might also be possible to get the BCL team to adopt zlib-ng over the Intel zlib once it reaches release state.