@JimBobSquarePants 2020-04-08
@saucecontrol It might also be possible to get the BCL team to adopt zlib-ng over the Intel zlib once it reaches release state.

This is a hope of mine. I had a look at the project earlier looking for comparisons and it definitely seemed the best choice. Far more active than any other implementation.